refactored Chunk1 with bad expression tests and began Chunk 3

+++ SimpleCalculator/SimpleCalculator/Division.cs
@@ -10,7 +10,7 @@ namespace SimpleCalculator
     {
         public double Divide(int num1, int num2)
         {
-            double result = num1 / num2;
+            double result = (double)num1 /(double) num2;
             return result;
         }
     }
